+++ src/parseTools.js
@@ -1622,3 +1622,10 @@ function makeModuleReceiveWithVar(localName, moduleName, defaultValue, noAssert)
   }
   return ret;
 }
+
+function makeRemovedFSAssert(fsName) {
+  if (!ASSERTIONS) return;
+  var lower = fsName.toLowerCase();
+  if (SYSTEM_JS_LIBRARIES.indexOf('library_' + lower + '.js') >= 0) return '';
+  return "var " + fsName + " = '" + fsName + " is no longer included by default; build with -l" + lower + ".js';";
+}