+++ ChangeLog.md
@@ -17,6 +17,9 @@ See docs/process.md for how version tagging works.
 
 Current Trunk
 -------------
+ - Only MEMFS is included by default, others (NODEFS, IDBFS, WORKERFS, PROXYFS)
+   must be linked in explicitly, using `-lnodefs.js`, `-lidbfs.js`',
+   `-lworkerfs.js`, `-lproxyfs.js`. See #9645
 
 v1.39.0: 10/18/2019
 -------------------